pyproject.toml
[tool.poetry]
name = "audio-vectorize"
version = "0.1.0"
description = ""
authors = ["fahadsalamatrandhawa <dreamtwister61814@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-multipart = "^0.0.9"
speechrecognition = "^3.10.1"
pydub = "^0.25.1"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.16"
psycopg2 = "^2.9.9"
langchain-mongodb = "^0.1.3"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
fastapi = "^0.110.0"
langchain = "^0.1.13"
sentence-transformers = "^2.6.0"
langchain-community = "^0.0.29"
mypy = "^1.9.0"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
